+++ src/app/admin/admin-routing.module.ts
@@ -1,0 +1,29 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { AdminComponent, AdminDashboardComponent, ProductFormComponent } from './components';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: AdminComponent,
+    canActivate: [],
+    children: [
+      {
+        path: '',
+        canActivateChild: [],
+        children: [
+          { path: 'add', component: ProductFormComponent },
+          { path: 'edit/:productId', component: ProductFormComponent },
+          { path: '', component: AdminDashboardComponent }
+        ]
+      }
+    ]
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class AdminRoutingModule { }