+++ src/app/cart/components/checkout/checkout.component.ts
@@ -1,0 +1,21 @@
+import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+
+@Component({
+  selector: 'app-checkout',
+  templateUrl: './checkout.component.html',
+  styleUrls: ['./checkout.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class CheckoutComponent implements OnInit {
+  message: string;
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+  order(): void {
+    this.message = 'You have made an order. Thanks for choosing our service!';
+  }
+
+}