+++ src/app/common/common-routing.module.ts
@@ -1,0 +1,41 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule, PreloadAllModules, ExtraOptions } from '@angular/router';
+
+import { LoginComponent, PageNotFoundComponent } from './components';
+import { AuthGuard } from '../core';
+
+const extraOptions: ExtraOptions = {
+  preloadingStrategy: PreloadAllModules,
+  enableTracing: false
+};
+
+
+const routes: Routes = [
+  {
+    path: 'login',
+    component: LoginComponent,
+    data: { title: 'Login' }
+  },
+  {
+    path: 'admin',
+    canLoad: [AuthGuard],
+    loadChildren: 'app/admin/admin.module#AdminModule',
+    data: { title: 'Admin' }
+  },
+  {
+    path: '',
+    redirectTo: '/home',
+    pathMatch: 'full'
+  },
+  {
+    path: '**',
+    component: PageNotFoundComponent,
+    data: { title: 'Page Not Found' }
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forRoot(routes, extraOptions)],
+  exports: [RouterModule]
+})
+export class CommonRoutingModule {}