+++ src/app/common/components/index.ts
@@ -1,0 +1,4 @@
+export * from './footer/footer.component';
+export * from './header/header.component';
+export * from './login/login.component';
+export * from './page-not-found/page-not-found.component';