+++ src/app/components/index.ts
@@ -1,3 +1,0 @@
-export * from '../components/product/product.component';
-export * from  '../components/product-list/product-list.component';
-export * from  '../components/cart/cart.component';