+++ src/app/products/components/index.ts
@@ -1,0 +1,5 @@
+export * from './product-list/product-list.component';
+export * from './product/product.component';
+export * from './product-detail/product-detail.component';
+export * from './reviews/reviews.component';
+export * from './test-core-services/test-core-services.component';